refactor(81): extract AppStrings bundle from AppLocale.kt into AppStrings.kt
AppLocale.kt now contains only the AppLocale enum and LocalAppLocale
CompositionLocal. All string-related code (AppStrings data class,
EnStrings, FrStrings, appStrings(), and localizedName() extensions)
lives in the new AppStrings.kt, mirroring how AppTheme.kt holds just
the theme enum.
